+++ src/realm/table_view.hpp
@@ -316,14 +316,12 @@ public:
     // Sort m_row_indexes according to multiple columns
     void sort(std::vector<size_t> columns, std::vector<bool> ascending);
 
-    // Remove rows that are duplicated with respect to the column set passed as argument.
+    // Remove rows from returned TableView that are duplicated with respect to the column set passed as argument. 
     // distinct() will preserve the original order of the row pointers, also if the order is a result of sort()
-    // If two rows are indentical (for the given set of distinct-columns), then the last row is removed.
-    // You can call sync_if_needed() to update the distinct view, just like you can for a sorted view.
-    // Each time you call distinct() it will first fetch the full original TableView contents and then apply
-    // distinct() on that. So it distinct() does not filter the result of the previous distinct().
-    void distinct(size_t column);
-    void distinct(std::vector<size_t> columns);
+    // If two rows are indentical (for the given set of distinct-columns), then the last row is removed. You can
+    // call sync_if_needed() to update the distinct view, just like you can for a sorted view.
+    TableView distinct(size_t column);
+    TableView distinct(std::vector<size_t> columns);
 
     // Returns whether the rows are guaranteed to be in table order.
     // This is true only of unsorted TableViews created from either:
@@ -469,6 +467,8 @@ private:
     void adj_row_acc_erase_row(size_t row_ndx) noexcept;
     void adj_row_acc_move_over(size_t from_row_ndx, size_t to_row_ndx) noexcept;
     void adj_row_acc_clear() noexcept;
+
+    void distinct_internal(std::vector<size_t> columns);
 
     template<typename Tab>
     friend class BasicTableView;
@@ -875,8 +875,8 @@ inline TableViewBase::TableViewBase(const TableViewBase& tv):
     m_linked_row(tv.m_linked_row),
     m_linkview_source(tv.m_linkview_source),
     m_distinct_column_source(tv.m_distinct_column_source),
-    m_distinct_columns(std::move(tv.m_distinct_columns)),
-    m_sorting_predicate(std::move(tv.m_sorting_predicate)),
+    m_distinct_columns(tv.m_distinct_columns),
+    m_sorting_predicate(tv.m_sorting_predicate),
     m_auto_sort(tv.m_auto_sort),
     m_query(tv.m_query),
     m_start(tv.m_start),