+++ test/test_table_view.cpp
@@ -1871,155 +1871,107 @@ TEST(TableView_BacklinksWhenTargetRowMovedOrDeleted)
 TEST(TableView_Distinct)
 {
     // distinct() will preserve the original order of the row pointers, also if the order is a result of sort()
-    // If multiple rows are indentical for the given set of distinct-columns, then only the first is kept.
-    // You can call sync_if_needed() to update the distinct view, just like you can for a sorted view.
-    // Each time you call distinct() it will first fetch the full original TableView contents and then apply
-    // distinct() on that. So it distinct() does not filter the result of the previous distinct().
+    // If multiple rows are indentical for the given set of distinct-columns, then only the *first* is kept 
+    // (*not* the entry with lowest table row index, but the entry that occurs first in the TableView). You 
+    // can call sync_if_needed() to update the distinct view, just like you can for a sorted view. distinct()
+    // will return a new TableView and keep the original unchanged. 
+    // 
+    // NOTE: distinct() has, before public release, behaved such that it would first refetch the complete
+    // TableView, discarding any effects of previous distinct() calls, and then apply distinct on that full set
+    // of rows. This has now been changed into *not* forgetting the previous distinct() filters.
 
     // distinct() is internally based on the existing sort() method which is well tested. Hence it's not required
     // to test distinct() with all possible Realm data types.
 
     Table t;
-    t.add_column(type_String, "s", true);
-    t.add_column(type_Int, "i", true);
-    t.add_column(type_Float, "f", true);
+    t.add_column(type_String, "c0", true);
+    t.add_column(type_Int, "c1", true);
+    t.add_column(type_Float, "c2", true);
+    t.add_column(type_Int, "c3", true);
 
-    t.add_empty_row(7);
-    t.set_string(0, 0, StringData(""));
-    t.set_int(1, 0, 100);
+    t.add_empty_row(4);
+
+    t.set_string(0, 0, "foo");  // 0
+    t.set_int(1, 0, 1000000);
     t.set_float(2, 0, 100.);
+    t.set_int(3, 0, 111);
 
-    t.set_string(0, 1, realm::null());
-    t.set_int(1, 1, 200);
-    t.set_float(2, 1, 200.);
+    t.set_string(0, 1, "foo");  // 1
+    t.set_int(1, 1, 2000000);
+    t.set_float(2, 1, 100.);
+    t.set_int(3, 1, 111);
 
-    t.set_string(0, 2, StringData(""));
-    t.set_int(1, 2, 100);
-    t.set_float(2, 2, 100.);
+    t.set_string(0, 2, "bar");  // 2
+    t.set_int(1, 2, 3000000);
+    t.set_float(2, 2, 200.);
+    t.set_int(3, 2, 111);
 
-    t.set_string(0, 3, realm::null());
-    t.set_int(1, 3, 200);
-    t.set_float(2, 3, 200.);
-
-    t.set_string(0, 4, "foo");
-    t.set_int(1, 4, 300);
-    t.set_float(2, 4, 300.);
-
-    t.set_string(0, 5, "foo");
-    t.set_int(1, 5, 400);
-    t.set_float(2, 5, 400.);
-
-    t.set_string(0, 6, "bar");
-    t.set_int(1, 6, 500);
-    t.set_float(2, 6, 500.);
-
+    t.set_string(0, 3, "bar");  // 3
+    t.set_int(1, 3, 3000000);
+    t.set_float(2, 3, 300.);
+    t.set_int(3, 3, 222);
 
     TableView tv;
-    tv = t.where().find_all();
-    tv.distinct(0);
-    CHECK_EQUAL(tv.size(), 4);
-    CHECK_EQUAL(tv.get_source_ndx(0), 0);
-    CHECK_EQUAL(tv.get_source_ndx(1), 1);
-    CHECK_EQUAL(tv.get_source_ndx(2), 4);
-    CHECK_EQUAL(tv.get_source_ndx(3), 6);
+    TableView tv2;
 
+    // Simple distinct on [100., 100., 200., 300.]
+    tv = t.where().find_all();
+    tv = tv.distinct(2);
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_source_ndx(0), 0);
+    CHECK_EQUAL(tv.get_source_ndx(1), 2);
+    CHECK_EQUAL(tv.get_source_ndx(2), 3);
+
+    // Filter further on tv.
+    tv = tv.distinct(1);
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_source_ndx(0), 0);
+    CHECK_EQUAL(tv.get_source_ndx(1), 2);
+
+    // distinct() must preserve sorting order
     tv = t.where().find_all();
     tv.sort(0);
-    tv.distinct(0);
-    CHECK_EQUAL(tv.size(), 4);
-    CHECK_EQUAL(tv.get_source_ndx(0), 1);
+    // [bar, bar, foo, foo], stable sort, so table indexes = [2, 3, 0, 1]
+    tv = tv.distinct(0);
+    // [bar, foo], table indexes = [2, 0]
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_source_ndx(0), 2);
     CHECK_EQUAL(tv.get_source_ndx(1), 0);
-    CHECK_EQUAL(tv.get_source_ndx(2), 6);
-    CHECK_EQUAL(tv.get_source_ndx(3), 4);
 
+    // Now try distinct on int+float column
     tv = t.where().find_all();
-    tv.sort(0, false);
-    tv.distinct(std::vector<size_t>{0});
-    CHECK_EQUAL(tv.get_source_ndx(0), 4);
-    CHECK_EQUAL(tv.get_source_ndx(1), 6);
-    CHECK_EQUAL(tv.get_source_ndx(2), 0);
-    CHECK_EQUAL(tv.get_source_ndx(3), 1);
-
-    // Note here that our stable sort will sort the two "foo"s like row {4, 5}
-    tv = t.where().find_all();
-    tv.sort(0, false);
-    tv.distinct(std::vector<size_t>{0, 1});
-    CHECK_EQUAL(tv.size(), 5);
-    CHECK_EQUAL(tv.get_source_ndx(0), 4);
-    CHECK_EQUAL(tv.get_source_ndx(1), 5);
-    CHECK_EQUAL(tv.get_source_ndx(2), 6);
-    CHECK_EQUAL(tv.get_source_ndx(3), 0);
-    CHECK_EQUAL(tv.get_source_ndx(4), 1);
-
-
-    // Now try distinct on string+float column. The float column has the same values as the int column
-    // so the result should equal the test above
-    tv = t.where().find_all();
-    tv.sort(0, false);
-    tv.distinct(std::vector<size_t>{0, 1});
-    CHECK_EQUAL(tv.size(), 5);
-    CHECK_EQUAL(tv.get_source_ndx(0), 4);
-    CHECK_EQUAL(tv.get_source_ndx(1), 5);
-    CHECK_EQUAL(tv.get_source_ndx(2), 6);
-    CHECK_EQUAL(tv.get_source_ndx(3), 0);
-    CHECK_EQUAL(tv.get_source_ndx(4), 1);
-
+    tv = tv.distinct(std::vector<size_t>{2, 3});
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_source_ndx(0), 0);
+    CHECK_EQUAL(tv.get_source_ndx(1), 2);
+    CHECK_EQUAL(tv.get_source_ndx(2), 3);
 
     // Same as previous test, but with string column being Enum
     t.optimize(true); // true = enforce regardless if Realm thinks it pays off or not
     tv = t.where().find_all();
-    tv.sort(0, false);
-    tv.distinct(std::vector<size_t>{0, 1});
-    CHECK_EQUAL(tv.size(), 5);
-    CHECK_EQUAL(tv.get_source_ndx(0), 4);
-    CHECK_EQUAL(tv.get_source_ndx(1), 5);
-    CHECK_EQUAL(tv.get_source_ndx(2), 6);
-    CHECK_EQUAL(tv.get_source_ndx(3), 0);
-    CHECK_EQUAL(tv.get_source_ndx(4), 1);
-
+    tv = tv.distinct(std::vector<size_t>{2, 3});
+    CHECK_EQUAL(tv.size(), 3);
+    CHECK_EQUAL(tv.get_source_ndx(0), 0);
+    CHECK_EQUAL(tv.get_source_ndx(1), 2);
+    CHECK_EQUAL(tv.get_source_ndx(2), 3);
 
     // Now test sync_if_needed()
     tv = t.where().find_all();
-    // "", null, "", null, "foo", "foo", "bar"
-
-    tv.sort(0, false);
-    // "foo", "foo", "bar", "", "", null, null
-
-    CHECK_EQUAL(tv.size(), 7);
-    CHECK_EQUAL(tv.get_string(0, 0), "foo");
-    CHECK_EQUAL(tv.get_string(0, 1), "foo");
-    CHECK_EQUAL(tv.get_string(0, 2), "bar");
-    CHECK_EQUAL(tv.get_string(0, 3), "");
-    CHECK_EQUAL(tv.get_string(0, 4), "");
-    CHECK(tv.get_string(0, 5).is_null());
-    CHECK(tv.get_string(0, 6).is_null());
-
-    tv.distinct(std::vector<size_t>{0});
-    // "foo", "bar", "", null
-
-    t.remove(6); // remove "bar"
-    // access to tv undefined; may crash
-
-    tv.sync_if_needed();
-    // "foo", "", null
-
+    // Simple distinct on [100., 100., 200., 300.]
+    tv = tv.distinct(2);
     CHECK_EQUAL(tv.size(), 3);
-    CHECK_EQUAL(tv.get_string(0, 0), "foo");
-    CHECK_EQUAL(tv.get_string(0, 1), "");
-    CHECK(tv.get_string(0, 2).is_null());
+    CHECK_EQUAL(tv.get_source_ndx(0), 0);
+    CHECK_EQUAL(tv.get_source_ndx(1), 2);
+    CHECK_EQUAL(tv.get_source_ndx(2), 3);
 
-    // Remove distinct property by providing empty column list. Now TableView should look like it
-    // did just after our last tv.sort(0, false) above, but after having executed table.remove(6)
-    tv.distinct(std::vector<size_t>{});
-    // "foo", "foo", "", "", null, null
-    CHECK_EQUAL(tv.size(), 6);
-    CHECK_EQUAL(tv.get_string(0, 0), "foo");
-    CHECK_EQUAL(tv.get_string(0, 1), "foo");
-    CHECK_EQUAL(tv.get_string(0, 2), "");
-    CHECK_EQUAL(tv.get_string(0, 3), "");
-    CHECK(tv.get_string(0, 4).is_null());
-    CHECK(tv.get_string(0, 5).is_null());
+    t.remove(3);
+    // access to tv is now undefined; may crash. So we do a sync_if_needed()
+    tv.sync_if_needed();
+    CHECK_EQUAL(tv.size(), 2);
+    CHECK_EQUAL(tv.get_source_ndx(0), 0);
+    CHECK_EQUAL(tv.get_source_ndx(1), 2);
 }
+
 
 TEST(TableView_IsInTableOrder)
 {